Standard telecomm interface to eForth, here in the lab.
The other variant (next commit) slows down the upload
of code to the STM32F407 Discovery target board, slow
enough to verify that lines are resulting in an 'ok'
prompt instead of a '?' (question mark) response.

It's still a bit fast for that purpose, but much slower
than stock (this very commit).

	new file:   pp-picot.sh

On branch assy-exp-ae-console-