TURNKEY'd application - displays logo on Lumex in blue
Not sure why this sometimes did not result
in a bootable system.

Suspect:
4 ERASE_SECTOR (any number higher than 3)

Possibly: ran 'COLD' and 'HEX' unthinkingly.

Really don't know why the turnkey failed.

Now it works (with this very version) and
this version is fairly clean (no substituted
signon variants).

Should be good to go!

	new file:   eforth-v-7-20--plus-USART6-dec-20-2020-a.fs

On branch dvlp-aa-