Now resident on target vers b2c-j0-
26 Jan 18:44:52

Had one upload error, patched without reupload.

Mistook ceflogo for 'eflogo' - ceflogo was purged.

Possibility of crashing 'signon' since it had that
error, and then 'BOOT with it, so redid it and
rewrote the TURNKEY.  Seems to work. ;)

Takeaway: this very file (this commit) should be
fine - but keep an eye on it. ;)

	modified:   eforth-v-7-20--plus-USART6-jan-26-2021-b.fs

On branch dvlp-cc-