blink_until a keystroke is received
This allows the target to keep blinking when the
keyboard is idle. ON time hard-coded - OFF time
is a bit more flexible.

Semi-tested prior to cosmetic edits.

Prints (emits) the single acquired character from
the keyboard, and then exits.

Meant to give a sense that the target is powered on,
when not presently in use. ;)

A quick glance at the machine tells that story,
especially when it hasn't been used - at all -
in several hours.

	new file:   blink_until-aa.fs

On branch dvlp-cc-